コントローラの機能テストをtest/controllersに移動し、ファイル名とクラス名を統一
