+++ test/controllers/api/pub_sub_controller_test.rb
