--- test/integration/api/reading_circles_test.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-require 'test_helper'
-
-class API::ReadingCirclesTest < ActionDispatch::IntegrationTest
-  test 'GET /api/reading_circles.json' do
-    get api_reading_circles_path(format: :json)
-    assert_response :unauthorized
-
-    token = create_token('kimura', 'testtest')
-    get api_reading_circles_path(format: :json),
-        headers: { 'Authorization' => "Bearer #{token}" }
-    assert_response :ok
-  end
-end
